+++ examples/blast-with-force-and-simulate/2-run-blast.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 echo >> pipeline.log
-echo "In `basename $0`" >> pipeline.log
+echo "In $(basename $0)" >> pipeline.log
 echo "SP_FORCE is $SP_FORCE" >> pipeline.log
 echo "SP_SIMULATE is $SP_SIMULATE" >> pipeline.log