+++ examples/double-collect/2-category-emit.sh
@@ -2,7 +2,7 @@
 
 # Just emit a new task name for each category.
 
-for category in `cut -f1 -d' ' categories`
+for category in $(cut -f1 -d' ' categories)
 do
     # Log to a file just for this category, since all category jobs will
     # run in parallel.