+++ examples/double-collect/4-summarize.sh
@@ -1,13 +1,13 @@
 #!/bin/bash
 
 echo >> pipeline.log
-echo "Running `basename $0` $@" >> pipeline.log
+echo "Running $(basename $0) $@" >> pipeline.log
 
-for category in `cut -f1 -d' ' categories`
+for category in $(cut -f1 -d' ' categories)
 do
     echo "Category $category:"
     cat output/$category.result
     echo
 done > output/SUMMARY
 
-echo "Finished `basename $0`" >> pipeline.log
+echo "Finished $(basename $0)" >> pipeline.log